updating docker-compose.yml to create all the containers at one and deploy it in any docker orchestrator.
Dockerfile, all the projetcs has tha same dockerfile using a  aspnetcore-build:1.1.1 image

Permissions: there is a bug .NETCore on Unix based system that does not allow the connection at the host using DNS,  this afect the redis.. I implpemented a workarround for  this, but even with that the Distributed Cache is not working properly. I have to remove the cache from permissions in order to have it working by now. I'll create a issue for that.