dockerfile: I changed it using the same approach for others services at arda.
I also have tried something different to create a certificate during the container creation, but it was not necessary I kept it there for documetation and to highlight the possible usage of the dockerfile.

Program.cs - Changed how the service is registering the ports needed (80 and 443), it would be better to receive the port  by env variable, this way we can change it at the compose whitout change the arda.main container.